fix(editor): render file manager images inline
