--- .github/workflows/tmp-file-manager-image-inline.yml
@@ -1,56 +1,0 @@
-name: Apply file manager image inline patch
-
-on:
-  push:
-    branches:
-      - fix/v1.4.6-file-manager-image-inline
-    paths:
-      - .github/workflows/tmp-file-manager-image-inline.yml
-
-permissions:
-  contents: write
-
-jobs:
-  apply:
-    runs-on: ubuntu-24.04
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: fix/v1.4.6-file-manager-image-inline
-          fetch-depth: 0
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 22
-          cache: npm
-          cache-dependency-path: frontend/package-lock.json
-
-      - name: Apply implementation patch
-        run: python scripts/tmp-apply-file-manager-image-inline.py
-
-      - name: Install frontend dependencies
-        working-directory: frontend
-        run: npm ci
-
-      - name: Run focused tests
-        working-directory: frontend
-        run: npm run test:run -- src/lib/__tests__/existingAttachmentInsert.test.ts
-
-      - name: Type-check frontend
-        working-directory: frontend
-        run: npx tsc -b --pretty false
-
-      - name: Commit verified implementation
-        shell: bash
-        run: |
-          set -euo pipefail
-          git rm -f .github/workflows/tmp-file-manager-image-inline.yml
-          git rm -f scripts/tmp-apply-file-manager-image-inline.py
-          git config user.name "Drug"
-          git config user.email "cropflre2333@gmail.com"
-          git add frontend/src/components/TiptapEditor.tsx
-          git add frontend/src/components/MarkdownEditorImpl.tsx
-          git add frontend/src/lib/existingAttachmentInsert.ts
-          git add frontend/src/lib/__tests__/existingAttachmentInsert.test.ts
-          git commit -m "fix(editor): render file manager images inline"
-          git push origin HEAD:fix/v1.4.6-file-manager-image-inline
