+++ frontend/src/components/TiptapEditor.tsx
@@ -57,6 +57,7 @@ import { shouldEmitTitleUpdate, shouldSkipTitleChange, shouldSyncTitleValue } fr
 import { resolveEditorLifecycleSave } from "@/lib/editorLifecycleSafety";
 import { api, resolveAttachmentUrl } from "@/lib/api";
 import { uploadAndInsertImage } from "@/lib/imageUploadService";
+import { isInlineImageAttachment } from "@/lib/existingAttachmentInsert";
 import {
   buildReplacedImageAttrs,
   getImageCopySource,
@@ -4382,13 +4383,25 @@ const TiptapEditor = forwardRef<NoteEditorHandle, TiptapEditorProps>(function Ti
       toast.error(t("tiptap.attachmentInsertPositionLost", { defaultValue: "插入位置已失效，请重试" }));
       return;
     }
-    editor
-      .chain()
-      .focus()
-      .insertContent(buildAttachmentLinkHtml(item.filename, item.url, item.size))
-      .run();
+
+    const inlineImage = isInlineImageAttachment(item);
+    const chain = editor.chain().focus();
+    if (inlineImage) {
+      chain.setImage({
+        src: item.url,
+        alt: item.filename,
+        title: item.filename,
+      }).run();
+    } else {
+      chain
+        .insertContent(buildAttachmentLinkHtml(item.filename, item.url, item.size))
+        .run();
+    }
+
     closeAttachmentLibrary();
-    toast.success(t("tiptap.attachmentLinkInserted", { defaultValue: "附件链接已插入" }));
+    toast.success(inlineImage
+      ? t("tiptap.imageInsertedFromFileManager", { defaultValue: "图片已插入" })
+      : t("tiptap.attachmentLinkInserted", { defaultValue: "附件链接已插入" }));
   }, [closeAttachmentLibrary, editor, restoreEditorInsertAnchor, t]);
 
   const handleImageUpload = useCallback(() => {
